Add RBAC enhancement interfaces for V2
